+++ lib/src/models/page_status.dart
@@ -16,13 +16,22 @@ class PageStatus extends Equatable {
   /// Set if this output was was captured from a [PopupRoute]
   final bool isPopup;
 
+  /// Future that ends with the route transition animation
+  final Future transitionEnded;
+
+  /// Duration being waited for by [transitionEnded]
+  @visibleForTesting
+  final Duration transitionDuration;
+
   /// Construct a new page status
   PageStatus({
     String? name,
     this.disabled = false,
     this.isPopup = false,
+    this.transitionDuration = Duration.zero,
   })  : name = name ?? '${DateTime.now()}',
-        nameMissing = name == null;
+        nameMissing = name == null,
+        transitionEnded = Future.delayed(transitionDuration);
 
   @override
   List<Object?> get props => [name, nameMissing, disabled, isPopup];